Eliminar gasto de una factura dividida: el borrado fallaba en silencio
Causa real: al borrar UNA línea de una factura con varias (mismo `grupo`),
el UPDATE que recalcula el moviban compartido hacía
"ingreso = CASE WHEN ingreso > 0 THEN $1 ELSE 0 END" sin castear $1. Sin
el cast, Postgres infiere el tipo del parámetro por el otro lado del CASE
(el literal "0", integer) y el UPDATE revienta apenas el total recalculado
queda con decimales — que es lo normal al borrar una línea cualquiera.
Mismo bug ya corregido antes en la eliminación de transferencias
(piernaHermanaTransferencia); acá nunca se había tocado. Verificado con
un dry-run (BEGIN...ROLLBACK, sin tocar datos reales) reproduciendo el
error exacto y confirmando que $1::numeric lo resuelve.

El backend SÍ devolvía el error (500), pero:
- GestionGastosTable.vue lo mandaba solo a console.error, nunca a la
  pantalla — de ahí "no me los elimina ni me muestra error". Se agrega
  un v-snackbar con el mensaje real del backend.
- El store leía err.response.data.message en actualizarGasto/eliminarGasto,
  pero el backend siempre manda `error`, no `message` — el fallback
  genérico tapaba el mensaje real incluso donde sí se mostraba.

De paso: nuevo script inventario-vue/scripts/dev-local.js + npm run
dev:local, para poder levantar el dev server de Vite apuntando al backend
LOCAL (VITE_API_TARGET) sin depender de la sintaxis "VAR=valor" del shell,
que no funciona en Windows. Usado para reproducir y verificar este fix
end-to-end antes de comitear.

+++ server-modular-UNICO.js
@@ -12134,10 +12134,16 @@ app.delete('/api/contabilidad/gastos/:codigo', async (req, res) => {
             await client.query('DELETE FROM moviban WHERE gasto = $1 AND empresa = $2', [grupoKeyDel, empresa]);
         } else {
             const totalRestante = parseFloat(restantesRes.rows[0].total) || 0;
+            // $1::numeric: sin el cast, Postgres infiere el tipo del parámetro
+            // por el otro lado del CASE (el literal "0", que lee como integer)
+            // y el UPDATE revienta apenas el total recalculado queda con
+            // decimales — que es exactamente lo normal al borrar UNA línea de
+            // una factura de varias. Mismo bug que ya se había corregido en
+            // la eliminación de transferencias (piernaHermanaTransferencia).
             await client.query(
                 `UPDATE moviban
-                 SET ingreso = CASE WHEN ingreso > 0 THEN $1 ELSE 0 END,
-                     egreso  = CASE WHEN egreso  > 0 THEN $1 ELSE 0 END
+                 SET ingreso = CASE WHEN ingreso > 0 THEN $1::numeric ELSE 0 END,
+                     egreso  = CASE WHEN egreso  > 0 THEN $1::numeric ELSE 0 END
                  WHERE gasto = $2 AND empresa = $3`,
                 [totalRestante, grupoKeyDel, empresa]
             );
